Refactor PivotTable tests: eliminate duplicates and consolidate structure
- Deleted duplicate PivotTableCommandsTests.Realistic.cs from Commands/PivotTable/
- Moved PivotTableCommandsTests.OlapFields.cs to Integration/Commands/PivotTable/
- Removed Collection attribute (no longer needed - tests run independently)
- Simplified test names: DataModelPivotTable_* -> OlapPivot_* (clearer, shorter)
- Removed redundant assertions (kept essential binary assertions only)
- Removed empty Commands/PivotTable/ directory

Test organization now follows established pattern:
- All PivotTable tests in Integration/Commands/PivotTable/
- OlapFields tests standalone (not partial class - uses different fixture)
- Clear naming: AddRowField_OlapPivot_AddsFieldToRows
- Clean hierarchy: Integration -> Commands -> PivotTable

Build: 0 warnings, 0 errors
Tests: SortField_OlapPivot_SortsFieldSuccessfully PASSES (validates refactoring)
